requirements.txt
pyrogram
tgcrypto
datetime
pytz
python-dotenv